feat: add new statistics endpoints and update data models for bordereau and facture

+++ src/main/java/com/soosmart/facts/dto/stat/Chart.java
@@ -2,6 +2,6 @@ package com.soosmart.facts.dto.stat;
 
 public record Chart(
         String label,
-        Integer value
+        Long value
 ) {
 }
